refactor: 解耦 egui 依赖，提取 GraphController + AppState
- #78: serial.rs 用 egui::Pos2::new() 替换 eframe::egui::pos2()
- #77: ExecutionManager 用通用回调替换 egui::Context
- #75: 从 NodeViewer 提取框架无关图逻辑到 GraphController
- #67: WidgetRegistry 留在 UI 层，不在 GraphController 中
- #76: 从 App 提取业务逻辑到 AppState（后端/文件/FPS/主题）
