+++ crates/nodeimg-app/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod app;
+pub mod app_state;
 pub mod execution;
 pub mod gpu;
 pub mod node;
