+++ crates/nodeimg-app/src/node/mod.rs
@@ -1,3 +1,4 @@
+pub mod graph_controller;
 pub mod serial;
 pub mod viewer;
 pub mod widget;
